10/8 <model 파일 위치 변경> 이세희 android_prep.py : 파일위치를 model 폴더의 하위로 이동 OCRModel.py : 파일위치를 model 폴더의 하위로 이동, 클래스명과 파일명 수정 model -> OCRModel app.py : system PATH 를 프로젝트경로로 변경
